split payload `Signer` and `Attestor` into distinct files
