+++ internal/pkg/cosign/payload/attestor.go
@@ -28,18 +28,7 @@ import (
 	"github.com/sigstore/cosign/pkg/oci/static"
 	"github.com/sigstore/cosign/pkg/types"
 	"github.com/sigstore/sigstore/pkg/signature"
-	signatureoptions "github.com/sigstore/sigstore/pkg/signature/options"
 )
-
-type payloadSigner struct {
-	payloadSigner         signature.Signer
-	payloadSignerOpts     []signature.SignOption
-	publicKeyProviderOpts []signature.PublicKeyOption
-
-	certPEM, chainPEM []byte
-}
-
-var _ cosign.Signer = (*payloadSigner)(nil)
 
 type payloadAttestor struct {
 	payloadSigner
@@ -48,50 +37,6 @@ type payloadAttestor struct {
 }
 
 var _ cosign.Attestor = (*payloadAttestor)(nil)
-
-func (ps *payloadSigner) signPayload(ctx context.Context, payload io.Reader) (payloadBytes, sig []byte, pk crypto.PublicKey, err error) {
-	payloadBytes, err = io.ReadAll(payload)
-	if err != nil {
-		return nil, nil, nil, err
-	}
-
-	sOpts := []signature.SignOption{signatureoptions.WithContext(ctx)}
-	sOpts = append(sOpts, ps.payloadSignerOpts...)
-	sig, err = ps.payloadSigner.SignMessage(bytes.NewReader(payloadBytes), sOpts...)
-	if err != nil {
-		return nil, nil, nil, err
-	}
-
-	pkOpts := []signature.PublicKeyOption{signatureoptions.WithContext(ctx)}
-	pkOpts = append(pkOpts, ps.publicKeyProviderOpts...)
-	pk, err = ps.payloadSigner.PublicKey(pkOpts...)
-	if err != nil {
-		return nil, nil, nil, err
-	}
-
-	return payloadBytes, sig, pk, nil
-}
-
-// Sign implements `Signer`
-func (ps *payloadSigner) Sign(ctx context.Context, payload io.Reader) (oci.Signature, crypto.PublicKey, error) {
-	payloadBytes, sig, pk, err := ps.signPayload(ctx, payload)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	b64sig := base64.StdEncoding.EncodeToString(sig)
-
-	var opts []static.Option
-	if len(ps.certPEM) > 0 {
-		opts = []static.Option{static.WithCertChain(ps.certPEM, ps.chainPEM)}
-	}
-	ociSig, err := static.NewSignature(payloadBytes, b64sig, opts...)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	return ociSig, pk, nil
-}
 
 // Attest implements `cosign.Attestor`
 func (pa *payloadAttestor) Attest(ctx context.Context, payload io.Reader) (oci.Signature, crypto.PublicKey, error) {
@@ -129,30 +74,6 @@ func (pa *payloadAttestor) Attest(ctx context.Context, payload io.Reader) (oci.S
 	}
 
 	return att, pk, nil
-}
-
-func newSigner(s signature.Signer,
-	sOpts []signature.SignOption,
-	pkOpts []signature.PublicKeyOption,
-	certPEM, chainPEM []byte) payloadSigner {
-	return payloadSigner{
-		payloadSigner:         s,
-		payloadSignerOpts:     sOpts,
-		publicKeyProviderOpts: pkOpts,
-
-		certPEM:  certPEM,
-		chainPEM: chainPEM,
-	}
-}
-
-// NewSigner returns a `cosign.Signer` which uses the given `signature.Signer` to sign requested payloads.
-// The cert and chain, if provided, will be included in returned `oci.Signature`s.
-func NewSigner(s signature.Signer,
-	sOpts []signature.SignOption,
-	pkOpts []signature.PublicKeyOption,
-	certPEM, chainPEM []byte) cosign.Signer {
-	ps := newSigner(s, sOpts, pkOpts, certPEM, chainPEM)
-	return &ps
 }
 
 // NewDSSEAttestor returns a `cosign.Attestor` which uses the given `signature.Signer` to create an attestation out of given payloads.
